- separates image viewer utility into its own library in src/lib/nmImageViewer - rewrote registration code in src/app/nano/lib/nmReg so that it can be   optionally run in a separate program using vrpn to communicate - the registration server program can be compiled in src/app/registration - some fiddling with the order of code to try to figure out a bug which   is causing texture images not to update even though everything appears to - separates image viewer utility into its own library in src/lib/nmImageViewer - rewrote registration code in src/app/nano/lib/nmReg so that it can be   optionally run in a separate program using vrpn to communicate - the registration server program can be compiled in src/app/registration - some fiddling with the order of code to try to figure out a bug which   is causing texture images not to update even though everything appears to   execute without error or memory problems - added some initialization of textures to nmGraphics to try to fix this problem - added getWorldToImageTransform() function to nmb_Image class so this   can be composed with result from registration server which only operates   on pixels and doesn't know anything about x,y or z units - makes it so that the create_closing_toplevel_with_notify not only sets   a variable to indicate whether or not the window is open but also makes it so   that if the C-code changes the value of the variable,   it causes the window to become visible or hidden depending on the value set - added some binary files for the sem edax libraries (I think these have been   flagged so that cvs won't try to merge them or something like that)

+++ src/lib/nmBase/nmb_Device.C
@@ -1,4 +1,7 @@
 #include <assert.h>
+#include <string.h>
+#include <strings.h>
+
 #include "nmb_Device.h"
 
 char * nmb_Device::encode_RequestSynchronization (long *len,
@@ -15,7 +18,7 @@ char * nmb_Device::encode_RequestSynchronization (long *len,
   if (comment) {
     comment_length = strlen(comment) + 1;
   }
-  *len = 3*sizeof(vrpn_int32) + comment_length;
+  *len = 3l*(long)sizeof(vrpn_int32) + comment_length;
 
   msgbuf = new char [*len];
   if (!msgbuf) {
